requirements.txt
acme==1.1.0
attrs==19.3.0
Automat==0.8.0
backports.entry-points-selectable==1.1.1
blinker==1.4
certbot==0.40.0
certbot-nginx==0.40.0
certifi==2019.11.28
chardet==3.0.4
click==8.0.3
colorama==0.4.3
ConfigArgParse==0.13.0
configobj==5.0.6
constantly==15.1.0
dbus-python==1.2.16
distlib==0.3.4
entrypoints==0.3
Flask==2.0.2
future==0.18.2
httplib2==0.14.0
hyperlink==19.0.0
idna==2.8
importlib-metadata==1.5.0
incremental==16.10.1
itsdangerous==2.0.1
Jinja2==3.0.3
josepy==1.2.0
jsonpatch==1.22
jsonpointer==2.0
jsonschema==3.2.0
keyring==18.0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
MarkupSafe==2.0.1
mock==3.0.5
more-itertools==4.2.0
netifaces==0.10.4
oauthlib==3.1.0
parsedatetime==2.4
pbr==5.4.5
pexpect==4.6.0
platformdirs==2.4.0
pyinotify==0.9.6
pymacaroons==0.13.0
pytz==2019.3
requests==2.22.0
requests-toolbelt==0.8.0
requests-unixsocket==0.2.0
suds-py3==1.4.5.0

